requirements.txt
openai>=1.0.0
flask>=2.3.0
pillow>=10.0.0
python-dotenv>=1.0.0